+++ quantify_scheduler/json_utils.py
@@ -37,7 +37,7 @@ class ScheduleJSONDecoder(json.JSONDecoder):
     The Quantify Schedule JSONDecoder.
 
     The ScheduleJSONDecoder is used to convert a string with JSON content into a
-    :class:`~quantify_scheduler.types.Schedule`.
+    :class:`quantify_scheduler.schedules.schedule.Schedule`.
 
     To avoid the execution of malicious code ScheduleJSONDecoder uses
     :func:`ast.literal_eval` instead of :func:`eval` to convert the data to an instance
@@ -52,9 +52,9 @@ class ScheduleJSONDecoder(json.JSONDecoder):
 
         The list of serializable classes can be extended with custom classes by
         providing the `modules` keyword argument. These classes have to implement
-        :class:`~quantify_scheduler.types.Operation` and overload the :code:`__str__`
-        and :code:`__repr__` methods in order to serialize and deserialize domain
-        objects into a valid JSON-format.
+        :class:`quantify_scheduler.operations.operation.Operation` and overload the
+        :code:`__str__` and :code:`__repr__` methods in order to serialize and
+        deserialize domain objects into a valid JSON-format.
 
         Keyword Arguments
         -----------------
@@ -71,12 +71,13 @@ class ScheduleJSONDecoder(json.JSONDecoder):
         )
 
         # Use local import to void Error('Operation' from partially initialized module
-        # 'quantify_scheduler.types')
-        from quantify_scheduler import (  # pylint: disable=import-outside-toplevel
+        # 'quantify_scheduler')
+        # pylint: disable=import-outside-toplevel
+        from quantify_scheduler import resources
+        from quantify_scheduler.operations import (  # pylint: disable=import-outside-toplevel
             acquisition_library,
             gate_library,
             pulse_library,
-            resources,
         )
 
         self._modules: List[ModuleType] = [
@@ -181,13 +182,13 @@ class ScheduleJSONEncoder(json.JSONEncoder):
         object.
         """
         # Use local import to void Error('Operation' from partially initialized module
-        # 'quantify_scheduler.types')
+        # 'quantify_scheduler')
         from quantify_scheduler import (  # pylint: disable=import-outside-toplevel
+            Operation,
             resources,
-            types,
         )
 
-        if isinstance(o, (types.Operation, resources.Resource)):
+        if isinstance(o, (Operation, resources.Resource)):
             return repr(o)
         if hasattr(o, "__dict__"):
             return o.__dict__