+++ quantify_scheduler/schedules/timedomain_schedules.py
@@ -9,12 +9,16 @@ from typing import List, Union
 import numpy as np
 from typing_extensions import Literal
 
-from quantify_scheduler.acquisition_library import SSBIntegrationComplex
+from quantify_scheduler import Schedule
 from quantify_scheduler.enums import BinMode
-from quantify_scheduler.gate_library import X90, Measure, Reset, Rxy, X
-from quantify_scheduler.pulse_library import DRAGPulse, IdlePulse, SquarePulse
+from quantify_scheduler.operations.acquisition_library import SSBIntegrationComplex
+from quantify_scheduler.operations.gate_library import X90, Measure, Reset, Rxy, X
+from quantify_scheduler.operations.pulse_library import (
+    DRAGPulse,
+    IdlePulse,
+    SquarePulse,
+)
 from quantify_scheduler.resources import ClockResource
-from quantify_scheduler.types import Schedule
 
 
 # pylint: disable=too-many-arguments