+++ tests/scheduler/test_acquisition_library.py
@@ -10,15 +10,15 @@ from unittest import TestCase
 import numpy as np
 import pytest
 
-from quantify_scheduler.acquisition_library import (
+from quantify_scheduler import Operation
+from quantify_scheduler.enums import BinMode
+from quantify_scheduler.operations.acquisition_library import (
     NumericalWeightedIntegrationComplex,
     SSBIntegrationComplex,
     Trace,
 )
-from quantify_scheduler.enums import BinMode
-from quantify_scheduler.gate_library import X90
-from quantify_scheduler.pulse_library import DRAGPulse
-from quantify_scheduler.types import Operation
+from quantify_scheduler.operations.gate_library import X90
+from quantify_scheduler.operations.pulse_library import DRAGPulse
 
 
 def test_ssb_integration_complex():