+++ CHANGELOG.rst
@@ -7,11 +7,23 @@ Unreleased
 
 * Visualization - resolved a bug where a schedule was modified when drawing a circuit diagram (#197, !250)
 * Visualization - made matplotlib schedule visualization methods accessible as methods `plot_circuit_diagram_mpl` and `plot_pulse_diagram_mpl` of the `Schedule` class (!253)
+* Bugfix - Ramp pulse sampling utilizing `np.linspace` behaviour changed. (!258)
+* Bugfix - Removed redundant `determine_absolute_timing` step in `qcompile`. (!259)
 
 Breaking changes
 ~~~~~~~~~~~~~~~~
 
 * InstrumentCoordinator - `last_schedule` is now a property (!252).
+* Structure - We have refactored the Operation and Schedule classes out of the types module and moved the different operation libraries (acquisition_library, gate_library, and pulse_library) (#217, !256).
+    * `quantify_scheduler.types.Operation` -> `quantify_scheduler.operations.operation.Operation`, the import `quantify_scheduler.Operation` still works.
+    * `quantify_scheduler.types.Schedule` -> `quantify_scheduler.schedules.schedule.Schedule`, the import `quantify_scheduler.Schedule` still works.
+    * `quantify_scheduler.types.CompiledSchedule` -> `quantify_scheduler.schedules.schedule.CompiledSchedule`
+    * `quantify_scheduler.types.ScheduleBase` -> `quantify_scheduler.schedules.schedule.ScheduleBase`
+    * `quantify_scheduler.types.AcquisitionMetadata` -> `quantify_scheduler.schedules.schedule.AcquisitionMetadata`
+    * `quantify_scheduler.acquisition_library` -> `quantify_scheduler.operations.acquisition_library`
+    * `quantify_scheduler.gate_library` -> `quantify_scheduler.operations.gate_library`
+    * `quantify_scheduler.pulse_library` -> `quantify_scheduler.operations.pulse_library`
+
 
 Merged branches and closed issues
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~